[CT-6-1] 통합 가능한 코드 통합
